Final requirements update before prompting Cursor:
Review @contributing.md  and @llm_prompts.md and @config.py and then review @README.md and update it with a plan to implement the API. If you need clarity on any requirements of the API in order to generate all the code, conduct an interview of me one point at a time to collect those details. We will review and finalize this plan before generating any code.